dont kill process because its manage by thread and not a process

+++ dbtunnel/chainlit.py
@@ -106,7 +106,8 @@ class ChainlitAppTunnel(DbTunnel):
         print("Starting chainlit...", flush=True)
 
         my_env = os.environ.copy()
-        subprocess.run(f"kill -9 $(lsof -t -i:{self._port})", capture_output=True, shell=True)
+        if self._share is True:
+            subprocess.run(f"kill -9 $(lsof -t -i:{self._port})", capture_output=True, shell=True)
 
         if self._share is False:
             cmd = ["chainlit", "run", self._chainlit_script_path, "-h", "--host", "0.0.0.0", "--port",
